Changed from NestedFSM to a StackFSM.
As I got to understand the problem in a deeper level, this decision
seemed to be the most suitable. NestedFSM is still a good solution, but
that approach would lead to a lot of code I don't want to maintain.

I decided to make a class for every state, so they can be distinguished
from one another using comparisson operations. I also added a StateRule
object that should make it easier to create rules that can be compared
to the targeted string (e.g. html or xml strings).